posture_backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
PyJWT==2.9.0
bcrypt==4.2.0
channels==4.1.0
python-dotenv==1.0.1
uvicorn==0.32.0
